// webapp/src/app/interface/tooltip.interface.ts
export interface ToolTips {
    chromosomeEncoding: string
    offspringSelection: string
    survivorSelection: string
    crossoverSelection: string
    mutationSelection: string
    populationSize: string
    mutationProbability: string
    crossoverProbability: string
    tournamentSize: string
    numberOfGenerations: string
    minimumParetoSetSize: string
    maximumParetoSetSize: string
    splitModulesWeight: string
    combineModulesWeight: string
    moveElementsBetweenModulesWeight: string
    conceptualModelType: string
    metaModelType: string
    dtdFile: string
    isUseCustomEdgeWeights: string
    edgeWeightTable: string
    numberOfElementsPerModule: string
    isUseWeightedSumMethod: string
    objectiveTable: string
    conceptualModelFile: string
}